Rename active comment interface
