+++ apps/frontend/src/app/comments/pipes/is-editing.pipe.ts
@@ -1,13 +1,13 @@
 import { Pipe, PipeTransform } from '@angular/core';
 import { Comment } from '../types/comment';
-import { ActiveCommentInterface, ActiveCommentType } from '../types/active-comment.interface';
+import { ActiveComment, ActiveCommentType } from '../types/active.comment';
 
 @Pipe({
   name: 'isEditing'
 })
 export class IsEditingPipe implements PipeTransform {
   // eslint-disable-next-line class-methods-use-this
-  transform(activeComment: ActiveCommentInterface | null, comment: Comment): boolean {
+  transform(activeComment: ActiveComment | null, comment: Comment): boolean {
     if (!activeComment) {
       return false;
     }
