+++ apps/frontend/src/app/comments/pipes/is-replying.pipe.ts
@@ -1,5 +1,5 @@
 import { Pipe, PipeTransform } from '@angular/core';
-import { ActiveCommentInterface, ActiveCommentType } from '../types/active-comment.interface';
+import { ActiveComment, ActiveCommentType } from '../types/active.comment';
 import { Comment } from '../types/comment';
 
 @Pipe({
@@ -7,7 +7,7 @@ import { Comment } from '../types/comment';
 })
 export class IsReplyingPipe implements PipeTransform {
   // eslint-disable-next-line class-methods-use-this
-  transform(activeComment: ActiveCommentInterface | null, comment: Comment): boolean {
+  transform(activeComment: ActiveComment | null, comment: Comment): boolean {
     if (!activeComment) {
       return false;
     }
